# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BenchBoard consultant availability and matching API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentProfile
      tags: [auth]
      summary: Get the signed-in user's workspace profile
      responses:
        "200":
          description: Current workspace profile, or an empty profile before onboarding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentProfile"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/role:
    post:
      operationId: createCurrentProfile
      tags: [auth]
      summary: Create the signed-in user's consultant or client profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileSetupInput"
      responses:
        "201":
          description: Created workspace profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentProfile"
        "400":
          description: Invalid setup details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List client projects
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a client project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a client project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/hire:
    post:
      operationId: hireConsultant
      tags: [projects]
      summary: Hire a consultant for a client project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HireInput"
      responses:
        "201":
          description: Created engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HireResponse"
        "400":
          description: Invalid hire request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project or consultant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a client project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project removed
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    CurrentProfile:
      type: object
      properties:
        role:
          type: ["string", "null"]
          enum: [consultant, client, null]
        profile:
          oneOf:
            - $ref: "#/components/schemas/Consultant"
            - $ref: "#/components/schemas/Client"
            - type: "null"
      required: [role, profile]
    ProfileSetupInput:
      type: object
      properties:
        role:
          type: string
          enum: [consultant, client]
        name:
          type: string
          minLength: 1
        title:
          type: string
        hourlyRate:
          type: number
          minimum: 0
        company:
          type: string
        industry:
          type: string
        bio:
          type: string
      required: [role, name]
...
    HireInput:
      type: object
      properties:
        consultantId:
          type: integer
          minimum: 1
      required: [consultantId]
    HireResponse:
      type: object
      properties:
        message:
          type: string
        engagement:
          type: object
          additionalProperties: true
      required: [message, engagement]
...
    AvailabilityStatus:
      type: string
      enum: [available, deployed, unavailable]
    Engagement:
      type: object
      properties:
        id:
          type: integer
        projectName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [id, projectName, startDate, endDate]
    Consultant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        skills:
          type: array
          items:
            type: string
        serviceOffers:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
        availabilityStatus:
          $ref: "#/components/schemas/AvailabilityStatus"
        engagement:
          type: ["object", "null"]
          $ref: "#/components/schemas/Engagement"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, title, skills, serviceOffers, hourlyRate, availabilityStatus, engagement, createdAt, updatedAt]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        industry:
          type: string
        bio:
          type: string
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, company, industry, bio, rating, createdAt, updatedAt]
...
    ProjectStatus:
      type: string
      enum: [planning, active, paused, completed]
    Project:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientId, name, description, skills, status, budget, startDate, endDate, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        clientId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 20
        skills:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
      required: [clientId, name, description, skills, status]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 20
        skills:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        budget:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]