# lib/api-zod/src/generated/types/currentProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BenchBoard consultant availability and matching API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from './client';
import type { Consultant } from './consultant';
import type { CurrentProfileRole } from './currentProfileRole';

export interface CurrentProfile {
  /** @nullable */
  role: CurrentProfileRole;
  profile: Consultant | Client | null;
}