# lib/api-zod/src/generated/types/hireInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BenchBoard consultant availability and matching API
 * OpenAPI spec version: 0.1.0
 */

export interface HireInput {
  /** @minimum 1 */
  consultantId: number;
}