# lib/api-zod/src/generated/types/hireResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BenchBoard consultant availability and matching API
 * OpenAPI spec version: 0.1.0
 */
import type { HireResponseEngagement } from './hireResponseEngagement';

export interface HireResponse {
  message: string;
  engagement: HireResponseEngagement;
}