# lib/api-zod/src/generated/types/profileSetupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BenchBoard consultant availability and matching API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileSetupInputRole } from './profileSetupInputRole';

export interface ProfileSetupInput {
  role: ProfileSetupInputRole;
  /** @minLength 1 */
  name: string;
  title?: string;
  /** @minimum 0 */
  hourlyRate?: number;
  company?: string;
  industry?: string;
  bio?: string;
}